Improve program <> partner flow
