+++ apps/web/app/app.dub.co/(dashboard)/[slug]/programs/[programId]/partners/add-partner-sheet.tsx
@@ -98,6 +98,7 @@ function AddPartnerSheetContent({ setIsOpen }: AddPartnerSheetProps) {
         url: program?.url,
         trackConversion: true,
         programId: program?.id,
+        folderId: program?.defaultFolderId,
       }),
     });
