+++ apps/web/lib/api/partners/enroll-partner.ts
@@ -128,6 +128,7 @@ export const enrollPartner = async ({
             programId: program.id,
             partnerId: upsertedPartner.id,
             folderId: program.defaultFolderId,
+            trackConversion: true,
           },
           include: includeTags,
         })
