+++ apps/web/lib/zod/schemas/programs.ts
@@ -6,7 +6,7 @@ import { LinkSchema } from "./links";
 import { RewardSchema } from "./rewards";
 import { parseDateSchema } from "./utils";
 
-export const HOLDING_PERIOD_DAYS = [0, 30, 60, 90];
+export const HOLDING_PERIOD_DAYS = [0, 14, 30, 60, 90];
 
 export const ProgramSchema = z.object({
   id: z.string(),
