+++ apps/web/ui/partners/partner-details-sheet.tsx
@@ -359,6 +359,8 @@ function PartnerApproval({
         key: shortKey,
         url: program?.url,
         trackConversion: true,
+        programId: program?.id,
+        folderId: program?.defaultFolderId,
       }),
     });
